api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6

SQLAlchemy==2.0.34
psycopg[binary]==3.2.1

# Pydantic with EmailStr support
pydantic[email]==2.8.2
email-validator==2.2.0

# Auth + hashing
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
